Database/seed/requirements.txt
mysql-connector-python==9.0.0
orjson==3.10.7
